# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KHU:DArchive API",
    "description": "APIs for storing past career records as a RAG-ready archive.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/documents/text": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create a text source document",
        "operationId": "create_text_document_api_documents_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextDocumentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextDocumentCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/documents/pdf": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload a PDF source document and optionally process it into experiences",
        "operationId": "create_pdf_document_api_documents_pdf_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_pdf_document_api_documents_pdf_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfDocumentCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/documents/{document_id}/process": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Process a source document into experiences",
        "operationId": "process_document_api_documents__document_id__process_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentProcessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/documents/{document_id}/processing-result": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Read document processing results",
        "operationId": "get_processing_result_api_documents__document_id__processing_result_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentProcessingResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_create_pdf_document_api_documents_pdf_post": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "process_document": {
            "type": "boolean",
            "title": "Process Document",
            "default": true
          },
          "metadata": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "file"
        ],
        "title": "Body_create_pdf_document_api_documents_pdf_post"
      },
      "DocumentProcessResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "experience_count": {
            "type": "integer",
            "title": "Experience Count"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "status",
          "experience_count",
          "question_count"
        ],
        "title": "DocumentProcessResponse"
      },
      "DocumentProcessingResultResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceProcessingSummary"
            },
            "type": "array",
            "title": "Experiences"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "status",
          "experiences"
        ],
        "title": "DocumentProcessingResultResponse"
      },
...
      "ExperienceProcessingSummary": {
        "properties": {
          "experience_id": {
            "type": "string",
            "title": "Experience Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "completeness_score": {
            "type": "number",
            "title": "Completeness Score"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionSummary"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "experience_id",
          "title",
          "completeness_score",
          "confidence_score"
        ],
        "title": "ExperienceProcessingSummary"
      },
...
      "PdfDocumentCreateResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "experience_count": {
            "type": "integer",
            "title": "Experience Count",
            "default": 0
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count",
            "default": 0
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceProcessingSummary"
            },
            "type": "array",
            "title": "Experiences"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "status"
        ],
        "title": "PdfDocumentCreateResponse"
      },
...
      "QuestionSummary": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "question_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Type"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "question",
          "priority",
          "status"
        ],
        "title": "QuestionSummary"
      },
...
      "TextDocumentCreateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "source_type",
          "text"
        ],
        "title": "TextDocumentCreateRequest"
      },
      "TextDocumentCreateResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "status"
        ],
        "title": "TextDocumentCreateResponse"
      }
    }
  }
}